Skip file names that are empty strings.
Empty strings may be necessary as placeholders, for example when a
plugin bundle wants to pass a value indicating "no files", but
Config::MVP doesn't allow an empty array to be passed. Note that only
the empty string (i.e. the string with zero characters) is skipped. You
could still specify a file named with only spaces and tabs if you really
wanted to.